Set position_argument explicitly
Changes geom_bar and label position back to pos so it can be
controlled explicitly with a positon adjustment (e.g. position_dodge
or position_stack). The old behaviour of just giving it one of "dodge"
or "stack" does not appear to be working anymore, as geom_label will
complain about missing width.